+++ XcodeML-Exc-Tools/src/exc/object/StructType.java
@@ -10,13 +10,13 @@ public class StructType extends CompositeType
     protected XobjList fTypeParams;
     protected Xobject finalProcedure;
 
-    public StructType(String id, String parent_id, XobjList id_list, int typeQualFlags, Xobject gccAttrs,
+    public StructType(String id, String parent_id, XobjList id_list, long typeQualFlags, Xobject gccAttrs,
                       Xobject[] codimensions)
     {
         super(Xtype.STRUCT, id, parent_id, id_list, typeQualFlags, gccAttrs, codimensions);
     }
 
-    public StructType(String id, String parent_id, XobjList id_list, XobjList proc_list, int typeQualFlags,
+    public StructType(String id, String parent_id, XobjList id_list, XobjList proc_list, long typeQualFlags,
 		      Xobject gccAttrs, XobjList typeParams, Xobject finalProcedure)
     {
         this(id, parent_id, id_list, typeQualFlags, gccAttrs, (Xobject[])null);
@@ -26,7 +26,7 @@ public class StructType extends CompositeType
     }
 
     public StructType(String id, boolean is_class, XobjString tag_names, 
-                                 XobjList id_list, int typeQualFlags, Xobject gccAttrs)
+                                 XobjList id_list, long typeQualFlags, Xobject gccAttrs)
     {
         super(Xtype.STRUCT, id, null, tag_names, id_list, typeQualFlags, gccAttrs, null);
         this.is_class = is_class;
